requirements.txt
pillow
transformers
torch
typeguard
torchvision
accelerate
segment_anything
opencv-python
google-generativeai
